Various timing changes to reducing passing about of values. Counter fixes for the KH eye issue, Shadow Hearts hang, possibly Grandia 3 hang, maybe others too. (and proper FFX loop fix, instead of my original hack :P)

+++ pcsx2/PsxCounters.h
@@ -20,7 +20,7 @@
 #define __PSXCOUNTERS_H__
 
 typedef struct {
-	u32 count, mode, target;
+	u64 count, mode, target;
 	u32 rate, interrupt, otarget;
 	u32 sCycle, Cycle;
 	u32 sCycleT, CycleT;
